* src/preproc/html/pre-html.cc (makeTempFiles): Activate new code, removing the old one. * src/utils/indxbib/indxbib.cc (main): Remove compiler warning.
* src/libs/libgroff/tmpfile.cc (xtmpfile): Fix guard for `namep'.

+++ src/utils/indxbib/indxbib.cc
@@ -229,7 +229,7 @@ int main(int argc, char **argv)
     name_max = file_name_max(".");
   const char *filename = p ? p + 1 : basename;
   if (name_max >= 0 &&
-      long(strlen(filename) + sizeof(INDEX_SUFFIX) - 1) > name_max)
+      (strlen(filename) + sizeof(INDEX_SUFFIX) - 1) > name_max)
     fatal("`%1.%2' is too long for a filename", filename, INDEX_SUFFIX);
   if (p) {
     p++;
